Move fetch_config resource to platform cookbook
Move load_cluster_config to shared, it is used by slurm, scheduler-plugin and by multiple configuration steps (init, update, finalize).
