+++ cookbooks/aws-parallelcluster-platform/spec/unit/resources/fetch_config_spec.rb
